portfolio/requirements.txt
blinker==1.6.3
click==8.1.8
colorama==0.4.6
Flask==2.2.5
Flask-Mail==0.9.1
flask-sqlalchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.1.1
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.6
MarkupSafe==2.1.5
python-dotenv==0.21.1
SQLAlchemy==2.0.41
typing-extensions==4.7.1
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
